Make JS output configurable
Exporting compiled templates as CommonJS modules should not be the
only available option.

With this commit a config item sets how the compiled templates can be
merged to the final JS file. This option (`config.hoganjs.wrapper` in
the main config files of Brunch), is a function that takes path and
content and returns a string.